Fixed up some errors that might have blocked the previous commits from working. Probably not.
Made gravity static so I can implement it to other stuffs and clean up the code a bit.

+++ Project2D/Softbody.cpp
@@ -40,9 +40,9 @@ void Softbody::Build(PhysicsScene* scene, glm::vec2 position, float spacing, flo
 			bool endOjJ = column == numOfColumns - 1;
 			bool endOfI = row == numOfRows - 1;
 
-			Sphere* s22 = (!endOfI && !endOjJ) ? spheres[(row + 1) * numOfColumns + (column + 1)] : nullptr;
-			Sphere* s02 = !endOjJ ? spheres[(row - 1) * numOfColumns + (column + 1)] : nullptr;
-			Sphere* s20 = !endOfI ? spheres[(row + 1) * numOfColumns + (column - 1)] : nullptr;
+			Sphere* s22 = (!endOfI && !endOjJ) ? spheres[(row + 1) * (int)numOfColumns + (column + 1)] : nullptr;
+			Sphere* s02 = !endOjJ ? spheres[(row - 1) * (int)numOfColumns + (column + 1)] : nullptr;
+			Sphere* s20 = !endOfI ? spheres[(row + 1) * (int)numOfColumns + (column - 1)] : nullptr;
 
 			if (s11 && s01)
 				scene->AddActor(new Spring(s11, s01, damping, springForce));
